[refactor] Move sequence comparison helpers to their own module
