+++ src/_pytest/assertion/util.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 
 import collections.abc
 from collections.abc import Callable
-from collections.abc import Iterable
 from collections.abc import Mapping
 from collections.abc import Sequence
 import dataclasses
@@ -16,15 +15,15 @@ from unicodedata import normalize
 
 from _pytest import outcomes
 import _pytest._code
-from _pytest._io.pprint import PrettyPrinter
 from _pytest._io.saferepr import saferepr
 from _pytest._io.saferepr import saferepr_unlimited
 from _pytest.assertion._compare_mapping import _compare_eq_mapping
+from _pytest.assertion._compare_sequence import _compare_eq_iterable
+from _pytest.assertion._compare_sequence import _compare_eq_sequence
 from _pytest.assertion._compare_set import _compare_eq_set
 from _pytest.assertion._compare_set import SET_COMPARISON_FUNCTIONS
 from _pytest.assertion._typing import _HighlightFunc
 from _pytest.compat import assert_never
-from _pytest.compat import running_on_ci
 from _pytest.config import Config
 from _pytest.config import UsageError
 
@@ -402,95 +401,6 @@ def _diff_text(
             lexer="diff",
         ).splitlines()
     )
-    return explanation
-
-
-def _compare_eq_iterable(
-    left: Iterable[object],
-    right: Iterable[object],
-    highlighter: _HighlightFunc,
-    verbose: int = 0,
-) -> list[str]:
-    if verbose <= 0 and not running_on_ci():
-        return ["Use -v to get more diff"]
-    # dynamic import to speedup pytest
-    import difflib
-
-    left_formatting = PrettyPrinter().pformat(left).splitlines()
-    right_formatting = PrettyPrinter().pformat(right).splitlines()
-
-    explanation = ["", "Full diff:"]
-    # "right" is the expected base against which we compare "left",
-    # see https://github.com/pytest-dev/pytest/issues/3333
-    explanation.extend(
-        highlighter(
-            "\n".join(
-                line.rstrip()
-                for line in difflib.ndiff(right_formatting, left_formatting)
-            ),
-            lexer="diff",
-        ).splitlines()
-    )
-    return explanation
-
-
-def _compare_eq_sequence(
-    left: Sequence[object],
-    right: Sequence[object],
-    highlighter: _HighlightFunc,
-    verbose: int = 0,
-) -> list[str]:
-    comparing_bytes = isinstance(left, bytes) and isinstance(right, bytes)
-    explanation: list[str] = []
-    len_left = len(left)
-    len_right = len(right)
-    for i in range(min(len_left, len_right)):
-        if left[i] != right[i]:
-            if comparing_bytes:
-                # when comparing bytes, we want to see their ascii representation
-                # instead of their numeric values (#5260)
-                # using a slice gives us the ascii representation:
-                # >>> s = b'foo'
-                # >>> s[0]
-                # 102
-                # >>> s[0:1]
-                # b'f'
-                left_value: object = left[i : i + 1]
-                right_value: object = right[i : i + 1]
-            else:
-                left_value = left[i]
-                right_value = right[i]
-
-            explanation.append(
-                f"At index {i} diff:"
-                f" {highlighter(repr(left_value))} != {highlighter(repr(right_value))}"
-            )
-            break
-
-    if comparing_bytes:
-        # when comparing bytes, it doesn't help to show the "sides contain one or more
-        # items" longer explanation, so skip it
-
-        return explanation
-
-    len_diff = len_left - len_right
-    if len_diff:
-        if len_diff > 0:
-            dir_with_more = "Left"
-            extra = saferepr(left[len_right])
-        else:
-            len_diff = 0 - len_diff
-            dir_with_more = "Right"
-            extra = saferepr(right[len_left])
-
-        if len_diff == 1:
-            explanation += [
-                f"{dir_with_more} contains one more item: {highlighter(extra)}"
-            ]
-        else:
-            explanation += [
-                f"{dir_with_more} contains {len_diff} more items, first extra item: {highlighter(extra)}"
-            ]
     return explanation
